requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.8.1
annotated-types==0.6.0
anyio==3.7.1
APScheduler==3.10.1
asgiref==3.7.2
async-timeout==4.0.3
asyncpg==0.25.0
asyncstdlib==3.10.7
attrs==23.2.0
bcrypt==4.0.1
boto3==1.28.8
botocore==1.31.85
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==2.0.12
click==8.1.7
cryptography==42.0.5
dnspython==2.6.1
email-validator==2.0.0.post1
environs==9.5.0
exceptiongroup==1.2.0
fastapi==0.85.0
fastapi-pagination==0.12.0
flake8==6.0.0
frozenlist==1.4.1
greenlet==2.0.2
gunicorn==20.0.4
h11==0.14.0
httptools==0.6.1
idna==3.6
iniconfig==2.0.0
jmespath==1.0.1
loguru==0.6.0
mailjet-rest==1.3.4
makers==0.2.6
Mako==1.3.2
MarkupSafe==2.1.5
marshmallow==3.21.1
mccabe==0.7.0
multidict==6.0.5
numpy==1.21.2
packaging==24.0
passlib==1.7.4
pendulum==2.1.2
pluggy==1.4.0
prometheus-client==0.15.0
psutil==5.9.4
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.9.2
pydantic_core==2.16.3
pyflakes==3.0.1
PyJWT==2.6.0
pytest==7.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.6
pytz==2023.3
pytzdata==2020.1
PyYAML==6.0.1
requests==2.26.0
s3transfer==0.6.2
six==1.16.0
sniffio==1.3.1
sqlakeyset==2.0.1708907391
SQLAlchemy==2.0.6
starlette==0.20.4
tomli==2.0.1
typing_extensions==4.10.0
tzlocal==5.2
urllib3==1.26.18
uvicorn==0.17.6
uvloop==0.19.0
watchgod==0.8.2
websockets==12.0
yarl==1.9.4
